Hybrid analyst/requirements.txt
google-genai>=0.8.0
fastapi>=0.110.0
uvicorn>=0.29.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0